Move Ephemeris to separate file in preparation for adding last three message types
