+++ rtcm3/gps.go
@@ -147,51 +147,6 @@ func (msg Message1004) Time() time.Time {
 	return DF004(msg.Epoch)
 }
 
-// GPS Ephemerides
-type Message1019 struct {
-	AbstractMessage
-	SatelliteID  uint8  `struct:"uint8:6"`
-	WeekNumber   uint16 `struct:"uint16:10"`
-	SVAccuracy   uint8  `struct:"uint8:4"`
-	CodeOnL2     uint8  `struct:"uint8:2"`
-	IDOT         int16  `struct:"int16:14"`
-	IODE         uint8  `struct:"uint8"`
-	Toc          uint16 `struct:"uint16"`
-	Af2          int8   `struct:"int8"`
-	Af1          int16  `struct:"int16"`
-	Af0          int32  `struct:"int32:22"`
-	IODC         uint16 `struct:"uint16:10"`
-	Crs          int16  `struct:"int16"`
-	DeltaN       int16  `struct:"int16"`
-	M0           int32  `struct:"int32"`
-	Cuc          int16  `struct:"int16"`
-	Eccentricity uint32 `struct:"uint32"`
-	Cus          int16  `struct:"int16"`
-	SrA          uint32 `struct:"uint32"`
-	Toe          uint16 `struct:"uint16"`
-	Cic          int16  `struct:"int16"`
-	Omega0       int32  `struct:"int32"`
-	Cis          int16  `struct:"int16"`
-	I0           int32  `struct:"int32"`
-	Crc          int16  `struct:"int16"`
-	Perigee      int32  `struct:"int32"`
-	OmegaDot     int32  `struct:"int32:24"`
-	Tgd          int8   `struct:"int8"`
-	SVHealth     uint8  `struct:"uint8:6"`
-	L2PDataFlag  bool   `struct:"uint8:1,variantbool"`
-	FitInterval  bool   `struct:"uint8:1,variantbool"`
-}
-
-func DeserializeMessage1019(data []byte) (msg Message1019) {
-	restruct.Unpack(data, binary.BigEndian, &msg)
-	return msg
-}
-
-func (msg Message1019) Serialize() []byte {
-	data, _ := restruct.Pack(binary.BigEndian, &msg)
-	return data
-}
-
 // Network Auxiliary Station Data Message
 type Message1014 struct {
 	AbstractMessage
